Change output to have filter output
This changes the output from ccfx_output_conv (the repertoire output) to
give files in the filter output rather than ccfx input.  This is
important because the line numbers given in the output also are for diff
output.

Eventually, we'll want to give both numbers and paths for the original
diff files, but that will be for another day.